requirements.txt
django==4.1.5
Pillow==9.4.0
djangorestframework==3.14.0
psycopg2>=2.8
django-cors-headers==3.13.0
django-filter==22.1
python-dotenv==0.21.1
mysqlclient==2.1.1